Implement logic of GenomePointer
GenomePointer is meant to point to one four bit number in a genome.
Since a genome is a list of eight bit numbers, we need two structures
here:

One with the pointer to the current byte in the genome. And another one
to pick whether we are currently pointing to the lower or higher four
bits of that byte.