Create execution machine
I briefly thought about using some kind of Peekable structure instead,
but that would over complicate things at this point. The original
implementation is made in such a way that minimizes wasted memory (i.e.
making sure that the four bits instruction doesn't waste the rest of the
bits in a number). Since I'm trying to make a Rust version as true to
the original intent as possible and making a Peekable with such
conditions would be a hard (although interesting) challenge,  I decided
to go the easy route.

But again, it'd be interesting. So maybe at some point I implement it,
instead of having two pointers.