Implement loops
Very straight forward. There's a stack with return addresses, when we
arrive to a Loop instruction we push the current pointer into the stack
and start looping if the condition doesn't hold. When we arrive to the
REP instruction, if there's something in the stack we take it out. If
the condition doesn't hold again, we move back to the pointer we took
from the stack.

If the condition holds when arriving to the LOOP, then we just move
forward to the next RET instruction.