Implement get_neighbor
To do so, we need to add the concept of a pond. And add that concept to
the VM.

Also, to simplify operations, I made the CellId be the coordinates in
the pond instead of a unique number. This has some problems (i.e. two
different cells that are in the pond in different times but in the same
position will have same id), but we can refactor again in the future
when those problems start mattering again. For now, this is easier.