Add statistics
So far, only to count how many times an instruction has been executed.

I feel that the `Index` implementation is a bit of an overkill, but I
wanted to feel how it works because I needed something similar for
another project and I didn't know it was possible. It's nice, although
unnecessarily verbose when you need `IndexMut`.