Randomly pick a cell position
This will be used to pick which cell to execute and which cell to
replace.

Note that it generates only one integer, despite needing two random
numbers. This is because the original code reduces as much as possible
the need of getting new integers and I want to keep as much
compatibility as possible with that code.