Write report
I rewrote here fold in a weird maner. One could argue that I could do
something like `pond.grind.into_iter().flatten()` and then do `filter`
and `fold` as usual and that would be correct.

However, apparently `into_iter` takes ownership of `self`, which we
don't want. Also, in order to abstract that in a nice way, for example,
by creating a `get_cell_iterator` function, it would require `Cell` to
be sized, which apparently is not right now.

Because fixing all that seems more complicated than just creating an
easy function to handle that logic, I'd just do the second. At least for
now.